refactor(git-utils): moving the clone dir function
