+++ packages/amplication-git-utils/src/utils/clone-dir/index.ts
@@ -1,0 +1,1 @@
+export * from "./clone-dir";
